requirements.txt
discord.py>=2.4.0
swelist>=0.1.7
python-dotenv>=1.0.0
aiohttp>=3.9.0
